FutureGlobals {future}	R Documentation

A representation of a set of globals used with futures

Description

A representation of a set of globals used with futures

Usage

FutureGlobals(object = list(), resolved = FALSE, total_size = NA_real_, ...)

Arguments

object
A named list.

resolved
A logical indicating whether these globals have been scanned for and resolved futures or not.

total_size
The total size of all globals, if known.

...
Not used.

Details

This class extends the Globals class by adding attributes resolved and total_size.

Value

An object of class FutureGlobals.

[Package future version 1.21.0 Index]