MulticoreFuture-class {future}	R Documentation

An multicore future is a future whose value will be resolved asynchronously in a parallel process

Description

An multicore future is a future whose value will be resolved asynchronously in a parallel process

Usage

MulticoreFuture(
  expr = NULL,
  substitute = TRUE,
  envir = parent.frame(),
  globals = TRUE,
  lazy = FALSE,
  ...
)

Arguments

expr
An R expression.

substitute
If TRUE, argument expr is substitute():ed, otherwise not.

envir
The environment from where global objects should be identified.

globals
(optional) a logical, a character vector, or a named list to control how globals are handled. For details, see section 'Globals used by future expressions' in the help for future().

lazy
If FALSE (default), the future is resolved eagerly (starting immediately), otherwise not.

...
Additional named elements passed to Future().

Value

An object of class MulticoreFuture.

See Also

To evaluate an expression using "multicore future", see function multicore().

[Package future version 1.21.0 Index]