A multiprocess future is a future whose value will be resolved asynchronously in a parallel process

A multiprocess future is a future whose value will be resolved asynchronously in a parallel process

MultiprocessFuture(expr = NULL, substitute = TRUE, envir = parent.frame(), ...)

Arguments

expr

An R expression.

substitute

If TRUE, argument expr is substitute():ed, otherwise not.

envir

The environment from where global objects should be identified.

...

Additional named elements passed to Future().

Value

MultiprocessFuture() returns an object of class MultiprocessFuture.