UniprocessFuture-class {future}	R Documentation

An uniprocess future is a future whose value will be resolved synchronously in the current process

Description

An uniprocess future is a future whose value will be resolved synchronously in the current process

Usage

UniprocessFuture(expr = NULL, substitute = TRUE, envir = parent.frame(), ...)

Arguments

expr
An R expression.

substitute
If TRUE, argument expr is substitute():d, otherwise not.

envir
The environment from where global objects should be identified.

...
Additional named elements passed to Future().

Value

UniprocessFuture() returns an object of class UniprocessFuture.

[Package future version 1.70.0 Index]