backtrace {future}	R Documentation

Back trace the expressions evaluated when an error was caught

Description

Back trace the expressions evaluated when an error was caught

Usage

backtrace(future, envir = parent.frame(), ...)

Arguments

future
A future with a caught error.

envir
the environment where to locate the future.

...
Not used.

Value

A list with the future's call stack that led up to the error.

Examples

my_log <- function(x) log(x)
foo <- function(...) my_log(...)

f <- future({ foo("a") })
res <- tryCatch({
  v <- value(f)
}, error = function(ex) {
  t <- backtrace(f)
  print(t)
})

[Package future version 1.21.0 Index]