clusterExportSticky {future}	R Documentation

Export globals to the sticky-globals environment of the cluster nodes

Description

Export globals to the sticky-globals environment of the cluster nodes

Usage

clusterExportSticky(cl, globals)

Arguments

cl
(cluster) A cluster object as returned by parallel::makeCluster().

globals
(list) A named list of sticky globals to be exported.

Details

This requires that the future package is installed on the cluster nodes.

Value

(invisible; cluster) The cluster object.

[Package future version 1.21.0 Index]