%globals% {future}	R Documentation

Specify globals and packages for a future assignment

Description

Specify globals and packages for a future assignment

Usage

fassignment %globals% globals
fassignment %packages% packages

Arguments

fassignment
The future assignment, e.g. x %<-% { expr }.

globals
(optional) a logical, a character vector, or a named list to control how globals are handled. For details, see section 'Globals used by future expressions' in the help for future().

packages
(optional) a character vector specifying packages to be attached in the R environment evaluating the future.

[Package future version 1.21.0 Index]