Gets a journal of events for a future

Gets a journal of events for a future

journal(x, ...)

Arguments

x

A Future object.

...

Not used.

Value

A data frame with columns:

step (character string)

start (POSIXct)

at (difftime)

duration (difftime)

future_label (character string)

future_uuid (character string)

session_uuid (character string)

The data frame is sorted by the at time. Note that the timestamps for the evaluate step are based on the local time on the worker. The system clocks on the worker and the calling R system may be out of sync.

Examples

library(future)
journal <- future:::journal

## Enable journaling of futures
oopts <- options(future.journal = TRUE)

plan(multisession, workers = 2L)

t_start <- Sys.time()
fs <- lapply(1:3, FUN = function(x) future({ Sys.sleep(x); sqrt(x) }))
vs <- value(fs)
js <- lapply(fs, FUN = journal, baseline = t_start)
print(js)
#> [[1]]
#>              step               start               at          duration
#> 1          create 2022-06-03 09:49:43 0.000000000 secs 2.735138e-03 secs
#> 6          launch 2022-06-03 09:49:43 0.003916979 secs 3.490925e-03 secs
#> 2       getWorker 2022-06-03 09:49:43 0.004441261 secs 1.103878e-04 secs
#> 3     eraseWorker 2022-06-03 09:49:43 0.005077362 secs 4.222393e-04 secs
#> 4  attachPackages 2022-06-03 09:49:43 0.005884886 secs 2.813339e-05 secs
#> 5   exportGlobals 2022-06-03 09:49:43 0.006271124 secs 2.942085e-04 secs
#> 34       evaluate 2022-06-03 09:49:43 0.007735491 secs 1.001204e+00 secs
#> 7        resolved 2022-06-03 09:49:44 0.093212843 secs 1.081467e-02 secs
#> 8        resolved 2022-06-03 09:49:44 0.135631323 secs 1.106501e-02 secs
#> 9        resolved 2022-06-03 09:49:44 0.169208527 secs 1.126671e-02 secs
#> 10       resolved 2022-06-03 09:49:44 0.203523874 secs 1.062346e-02 secs
#> 11       resolved 2022-06-03 09:49:44 0.236396790 secs 1.096797e-02 secs
#> 12       resolved 2022-06-03 09:49:44 0.270513535 secs 1.095223e-02 secs
#> 13       resolved 2022-06-03 09:49:44 0.305023909 secs 1.094151e-02 secs
#> 14       resolved 2022-06-03 09:49:44 0.339016914 secs 1.099586e-02 secs
#> 15       resolved 2022-06-03 09:49:44 0.372618914 secs 1.098537e-02 secs
#> 16       resolved 2022-06-03 09:49:44 0.406848431 secs 1.115870e-02 secs
#> 17       resolved 2022-06-03 09:49:44 0.441190720 secs 1.065302e-02 secs
#> 18       resolved 2022-06-03 09:49:44 0.475308895 secs 1.092315e-02 secs
#> 19       resolved 2022-06-03 09:49:44 0.508996010 secs 1.087928e-02 secs
#> 20       resolved 2022-06-03 09:49:44 0.543626785 secs 1.093841e-02 secs
#> 21       resolved 2022-06-03 09:49:44 0.578070164 secs 1.062679e-02 secs
#> 22       resolved 2022-06-03 09:49:44 0.612597704 secs 1.096201e-02 secs
#> 23       resolved 2022-06-03 09:49:44 0.647565842 secs 1.092744e-02 secs
#> 24       resolved 2022-06-03 09:49:44 0.682641029 secs 1.053858e-02 secs
#> 25       resolved 2022-06-03 09:49:44 0.716808081 secs 1.069856e-02 secs
#> 26       resolved 2022-06-03 09:49:44 0.751734495 secs 1.097751e-02 secs
#> 27       resolved 2022-06-03 09:49:44 0.787537813 secs 1.102424e-02 secs
#> 28       resolved 2022-06-03 09:49:44 0.823037386 secs 1.088762e-02 secs
#> 29       resolved 2022-06-03 09:49:44 0.858742952 secs 1.084733e-02 secs
#> 30       resolved 2022-06-03 09:49:44 0.894774914 secs 1.131177e-02 secs
#> 31       resolved 2022-06-03 09:49:44 0.930825472 secs 1.086283e-02 secs
#> 32       resolved 2022-06-03 09:49:44 0.966988564 secs 1.095080e-02 secs
#> 35       resolved 2022-06-03 09:49:44 1.003190756 secs 9.150505e-03 secs
#> 33         gather 2022-06-03 09:49:44 1.010030746 secs 4.196167e-05 secs
#> 36       resolved 2022-06-03 09:49:44 1.017231226 secs 3.027916e-05 secs
#> 37        resolve 2022-06-03 09:49:44 1.018057108 secs 1.144409e-05 secs
#>    future_label                          future_uuid
#> 1          <NA> 9f6a1b8a-cbb7-b05f-3099-ed8ab7101600
#> 6          <NA> 9f6a1b8a-cbb7-b05f-3099-ed8ab7101600
#> 2          <NA> 9f6a1b8a-cbb7-b05f-3099-ed8ab7101600
#> 3          <NA> 9f6a1b8a-cbb7-b05f-3099-ed8ab7101600
#> 4          <NA> 9f6a1b8a-cbb7-b05f-3099-ed8ab7101600
#> 5          <NA> 9f6a1b8a-cbb7-b05f-3099-ed8ab7101600
#> 34         <NA> 9f6a1b8a-cbb7-b05f-3099-ed8ab7101600
#> 7          <NA> 9f6a1b8a-cbb7-b05f-3099-ed8ab7101600
#> 8          <NA> 9f6a1b8a-cbb7-b05f-3099-ed8ab7101600
#> 9          <NA> 9f6a1b8a-cbb7-b05f-3099-ed8ab7101600
#> 10         <NA> 9f6a1b8a-cbb7-b05f-3099-ed8ab7101600
#> 11         <NA> 9f6a1b8a-cbb7-b05f-3099-ed8ab7101600
#> 12         <NA> 9f6a1b8a-cbb7-b05f-3099-ed8ab7101600
#> 13         <NA> 9f6a1b8a-cbb7-b05f-3099-ed8ab7101600
#> 14         <NA> 9f6a1b8a-cbb7-b05f-3099-ed8ab7101600
#> 15         <NA> 9f6a1b8a-cbb7-b05f-3099-ed8ab7101600
#> 16         <NA> 9f6a1b8a-cbb7-b05f-3099-ed8ab7101600
#> 17         <NA> 9f6a1b8a-cbb7-b05f-3099-ed8ab7101600
#> 18         <NA> 9f6a1b8a-cbb7-b05f-3099-ed8ab7101600
#> 19         <NA> 9f6a1b8a-cbb7-b05f-3099-ed8ab7101600
#> 20         <NA> 9f6a1b8a-cbb7-b05f-3099-ed8ab7101600
#> 21         <NA> 9f6a1b8a-cbb7-b05f-3099-ed8ab7101600
#> 22         <NA> 9f6a1b8a-cbb7-b05f-3099-ed8ab7101600
#> 23         <NA> 9f6a1b8a-cbb7-b05f-3099-ed8ab7101600
#> 24         <NA> 9f6a1b8a-cbb7-b05f-3099-ed8ab7101600
#> 25         <NA> 9f6a1b8a-cbb7-b05f-3099-ed8ab7101600
#> 26         <NA> 9f6a1b8a-cbb7-b05f-3099-ed8ab7101600
#> 27         <NA> 9f6a1b8a-cbb7-b05f-3099-ed8ab7101600
#> 28         <NA> 9f6a1b8a-cbb7-b05f-3099-ed8ab7101600
#> 29         <NA> 9f6a1b8a-cbb7-b05f-3099-ed8ab7101600
#> 30         <NA> 9f6a1b8a-cbb7-b05f-3099-ed8ab7101600
#> 31         <NA> 9f6a1b8a-cbb7-b05f-3099-ed8ab7101600
#> 32         <NA> 9f6a1b8a-cbb7-b05f-3099-ed8ab7101600
#> 35         <NA> 9f6a1b8a-cbb7-b05f-3099-ed8ab7101600
#> 33         <NA> 9f6a1b8a-cbb7-b05f-3099-ed8ab7101600
#> 36         <NA> 9f6a1b8a-cbb7-b05f-3099-ed8ab7101600
#> 37         <NA> 9f6a1b8a-cbb7-b05f-3099-ed8ab7101600
#>                            session_uuid
#> 1  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 6  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 2  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 3  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 4  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 5  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 34 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 7  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 8  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 9  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 10 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 11 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 12 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 13 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 14 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 15 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 16 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 17 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 18 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 19 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 20 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 21 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 22 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 23 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 24 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 25 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 26 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 27 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 28 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 29 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 30 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 31 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 32 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 35 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 33 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 36 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 37 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 
#> [[2]]
#>              step               start               at          duration
#> 1          create 2022-06-03 09:49:43 0.000000000 secs 2.350330e-03 secs
#> 6          launch 2022-06-03 09:49:43 0.002999544 secs 6.800103e-02 secs
#> 2       getWorker 2022-06-03 09:49:43 0.003484488 secs 1.008511e-04 secs
#> 3     eraseWorker 2022-06-03 09:49:43 0.003941059 secs 6.470275e-02 secs
#> 4  attachPackages 2022-06-03 09:49:44 0.069245100 secs 3.981590e-05 secs
#> 5   exportGlobals 2022-06-03 09:49:44 0.069707870 secs 3.583431e-04 secs
#> 67       evaluate 2022-06-03 09:49:44 0.071174622 secs 2.027961e+00 secs
#> 7        resolved 2022-06-03 09:49:44 0.096800804 secs 1.104760e-02 secs
#> 8        resolved 2022-06-03 09:49:44 0.139356375 secs 1.088166e-02 secs
#> 9        resolved 2022-06-03 09:49:44 0.173312426 secs 1.108718e-02 secs
#> 10       resolved 2022-06-03 09:49:44 0.206814289 secs 1.059103e-02 secs
#> 11       resolved 2022-06-03 09:49:44 0.240169764 secs 1.082516e-02 secs
#> 12       resolved 2022-06-03 09:49:44 0.274227381 secs 1.122808e-02 secs
#> 13       resolved 2022-06-03 09:49:44 0.308659554 secs 1.087308e-02 secs
#> 14       resolved 2022-06-03 09:49:44 0.342604160 secs 1.078176e-02 secs
#> 15       resolved 2022-06-03 09:49:44 0.376198530 secs 1.091194e-02 secs
#> 16       resolved 2022-06-03 09:49:44 0.410680294 secs 1.071048e-02 secs
#> 17       resolved 2022-06-03 09:49:44 0.444500208 secs 1.089311e-02 secs
#> 18       resolved 2022-06-03 09:49:44 0.478826046 secs 1.055598e-02 secs
#> 19       resolved 2022-06-03 09:49:44 0.512512922 secs 1.093698e-02 secs
#> 20       resolved 2022-06-03 09:49:44 0.547157526 secs 1.090980e-02 secs
#> 21       resolved 2022-06-03 09:49:44 0.581302166 secs 1.090860e-02 secs
#> 22       resolved 2022-06-03 09:49:44 0.616153955 secs 1.092339e-02 secs
#> 23       resolved 2022-06-03 09:49:44 0.651086807 secs 1.090860e-02 secs
#> 24       resolved 2022-06-03 09:49:44 0.685781240 secs 1.061463e-02 secs
#> 25       resolved 2022-06-03 09:49:44 0.720241308 secs 1.072955e-02 secs
#> 26       resolved 2022-06-03 09:49:44 0.755438805 secs 1.104116e-02 secs
#> 27       resolved 2022-06-03 09:49:44 0.791293621 secs 1.092482e-02 secs
#> 28       resolved 2022-06-03 09:49:44 0.826799631 secs 1.087499e-02 secs
#> 29       resolved 2022-06-03 09:49:44 0.862407446 secs 1.075935e-02 secs
#> 30       resolved 2022-06-03 09:49:44 0.898773670 secs 1.062226e-02 secs
#> 31       resolved 2022-06-03 09:49:44 0.934352398 secs 1.099825e-02 secs
#> 32       resolved 2022-06-03 09:49:44 0.970542192 secs 1.091647e-02 secs
#> 33       resolved 2022-06-03 09:49:44 1.010659456 secs 1.069355e-02 secs
#> 34       resolved 2022-06-03 09:49:45 1.043810129 secs 1.101398e-02 secs
#> 35       resolved 2022-06-03 09:49:45 1.077106714 secs 1.093483e-02 secs
#> 36       resolved 2022-06-03 09:49:45 1.110194683 secs 1.428795e-02 secs
#> 37       resolved 2022-06-03 09:49:45 1.147034168 secs 1.073599e-02 secs
#> 38       resolved 2022-06-03 09:49:45 1.179761410 secs 1.091552e-02 secs
#> 39       resolved 2022-06-03 09:49:45 1.212399244 secs 1.091695e-02 secs
#> 40       resolved 2022-06-03 09:49:45 1.245385408 secs 1.093030e-02 secs
#> 41       resolved 2022-06-03 09:49:45 1.278199196 secs 1.091433e-02 secs
#> 42       resolved 2022-06-03 09:49:45 1.311254978 secs 1.091504e-02 secs
#> 43       resolved 2022-06-03 09:49:45 1.343984604 secs 1.091385e-02 secs
#> 44       resolved 2022-06-03 09:49:45 1.377019405 secs 1.091743e-02 secs
#> 45       resolved 2022-06-03 09:49:45 1.410107374 secs 1.091814e-02 secs
#> 46       resolved 2022-06-03 09:49:45 1.443197966 secs 1.082683e-02 secs
#> 47       resolved 2022-06-03 09:49:45 1.476211309 secs 1.068282e-02 secs
#> 48       resolved 2022-06-03 09:49:45 1.509004354 secs 1.093578e-02 secs
#> 49       resolved 2022-06-03 09:49:45 1.542071581 secs 1.091099e-02 secs
#> 50       resolved 2022-06-03 09:49:45 1.575140953 secs 1.063371e-02 secs
#> 51       resolved 2022-06-03 09:49:45 1.608101606 secs 1.108050e-02 secs
#> 52       resolved 2022-06-03 09:49:45 1.641379356 secs 1.076341e-02 secs
#> 53       resolved 2022-06-03 09:49:45 1.673871756 secs 1.058269e-02 secs
#> 54       resolved 2022-06-03 09:49:45 1.707369328 secs 1.108384e-02 secs
#> 55       resolved 2022-06-03 09:49:45 1.740395069 secs 1.125526e-02 secs
#> 56       resolved 2022-06-03 09:49:45 1.774118185 secs 1.104140e-02 secs
#> 57       resolved 2022-06-03 09:49:45 1.807431698 secs 1.060963e-02 secs
#> 58       resolved 2022-06-03 09:49:45 1.840515375 secs 1.099062e-02 secs
#> 59       resolved 2022-06-03 09:49:45 1.875931978 secs 1.116610e-02 secs
#> 60       resolved 2022-06-03 09:49:45 1.909169436 secs 1.146483e-02 secs
#> 61       resolved 2022-06-03 09:49:45 1.942955732 secs 1.069236e-02 secs
#> 62       resolved 2022-06-03 09:49:45 1.976077557 secs 1.135659e-02 secs
#> 63       resolved 2022-06-03 09:49:45 2.009528637 secs 1.125813e-02 secs
#> 64       resolved 2022-06-03 09:49:46 2.043418646 secs 1.106167e-02 secs
#> 65       resolved 2022-06-03 09:49:46 2.076651335 secs 1.101327e-02 secs
#> 68       resolved 2022-06-03 09:49:46 2.109516859 secs 2.555847e-03 secs
#> 66         gather 2022-06-03 09:49:46 2.110424042 secs 3.671646e-05 secs
#> 69        resolve 2022-06-03 09:49:46 2.112807989 secs 1.645088e-05 secs
#>    future_label                          future_uuid
#> 1          <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 6          <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 2          <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 3          <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 4          <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 5          <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 67         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 7          <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 8          <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 9          <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 10         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 11         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 12         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 13         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 14         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 15         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 16         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 17         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 18         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 19         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 20         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 21         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 22         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 23         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 24         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 25         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 26         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 27         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 28         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 29         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 30         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 31         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 32         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 33         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 34         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 35         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 36         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 37         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 38         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 39         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 40         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 41         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 42         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 43         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 44         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 45         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 46         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 47         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 48         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 49         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 50         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 51         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 52         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 53         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 54         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 55         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 56         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 57         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 58         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 59         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 60         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 61         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 62         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 63         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 64         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 65         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 68         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 66         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#> 69         <NA> 389c7e1b-b27f-aae4-b562-76befddbfc69
#>                            session_uuid
#> 1  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 6  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 2  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 3  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 4  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 5  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 67 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 7  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 8  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 9  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 10 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 11 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 12 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 13 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 14 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 15 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 16 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 17 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 18 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 19 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 20 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 21 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 22 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 23 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 24 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 25 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 26 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 27 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 28 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 29 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 30 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 31 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 32 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 33 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 34 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 35 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 36 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 37 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 38 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 39 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 40 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 41 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 42 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 43 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 44 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 45 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 46 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 47 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 48 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 49 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 50 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 51 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 52 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 53 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 54 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 55 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 56 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 57 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 58 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 59 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 60 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 61 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 62 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 63 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 64 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 65 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 68 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 66 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 69 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 
#> [[3]]
#>               step               start               at          duration
#> 1           create 2022-06-03 09:49:44 0.000000000 secs 2.424479e-03 secs
#> 6           launch 2022-06-03 09:49:44 0.003191710 secs 9.334276e-01 secs
#> 2        getWorker 2022-06-03 09:49:44 0.003765583 secs 9.300451e-01 secs
#> 3      eraseWorker 2022-06-03 09:49:44 0.934167385 secs 3.929138e-04 secs
#> 4   attachPackages 2022-06-03 09:49:44 0.934989452 secs 3.647804e-05 secs
#> 5    exportGlobals 2022-06-03 09:49:44 0.935443640 secs 3.416538e-04 secs
#> 7         resolved 2022-06-03 09:49:44 0.950345516 secs 1.091385e-02 secs
#> 130       evaluate 2022-06-03 09:49:45 0.978157759 secs 3.004101e+00 secs
#> 8         resolved 2022-06-03 09:49:45 0.983838081 secs 1.095390e-02 secs
#> 9         resolved 2022-06-03 09:49:45 1.017007113 secs 1.088166e-02 secs
#> 10        resolved 2022-06-03 09:49:45 1.053691864 secs 1.099324e-02 secs
#> 11        resolved 2022-06-03 09:49:45 1.086733580 secs 1.072121e-02 secs
#> 12        resolved 2022-06-03 09:49:45 1.119635344 secs 1.047421e-02 secs
#> 13        resolved 2022-06-03 09:49:45 1.152278900 secs 1.089501e-02 secs
#> 14        resolved 2022-06-03 09:49:45 1.185410023 secs 1.063871e-02 secs
#> 15        resolved 2022-06-03 09:49:45 1.218074083 secs 1.088881e-02 secs
#> 16        resolved 2022-06-03 09:49:45 1.251132727 secs 1.081157e-02 secs
#> 17        resolved 2022-06-03 09:49:45 1.283858299 secs 1.086617e-02 secs
#> 18        resolved 2022-06-03 09:49:45 1.316925764 secs 1.088715e-02 secs
#> 19        resolved 2022-06-03 09:49:45 1.349981785 secs 1.089263e-02 secs
#> 20        resolved 2022-06-03 09:49:45 1.382990360 secs 1.090217e-02 secs
#> 21        resolved 2022-06-03 09:49:45 1.415914297 secs 1.078892e-02 secs
#> 22        resolved 2022-06-03 09:49:45 1.448899031 secs 1.089644e-02 secs
#> 23        resolved 2022-06-03 09:49:45 1.481942415 secs 1.089692e-02 secs
#> 24        resolved 2022-06-03 09:49:45 1.514735937 secs 1.090312e-02 secs
#> 25        resolved 2022-06-03 09:49:45 1.548207045 secs 1.095700e-02 secs
#> 26        resolved 2022-06-03 09:49:45 1.581167936 secs 1.076865e-02 secs
#> 27        resolved 2022-06-03 09:49:45 1.613602161 secs 1.135159e-02 secs
#> 28        resolved 2022-06-03 09:49:45 1.647449732 secs 1.056647e-02 secs
#> 29        resolved 2022-06-03 09:49:45 1.680781364 secs 1.105905e-02 secs
#> 30        resolved 2022-06-03 09:49:45 1.714173794 secs 1.091218e-02 secs
#> 31        resolved 2022-06-03 09:49:45 1.747274160 secs 1.083350e-02 secs
#> 32        resolved 2022-06-03 09:49:45 1.781624556 secs 1.176095e-02 secs
#> 33        resolved 2022-06-03 09:49:45 1.816115379 secs 1.096416e-02 secs
#> 34        resolved 2022-06-03 09:49:45 1.849821568 secs 1.105785e-02 secs
#> 35        resolved 2022-06-03 09:49:45 1.882615566 secs 1.093459e-02 secs
#> 36        resolved 2022-06-03 09:49:45 1.916504860 secs 1.096511e-02 secs
#> 37        resolved 2022-06-03 09:49:45 1.949975491 secs 1.110959e-02 secs
#> 38        resolved 2022-06-03 09:49:46 1.983531952 secs 1.070166e-02 secs
#> 39        resolved 2022-06-03 09:49:46 2.016677380 secs 1.052833e-02 secs
#> 40        resolved 2022-06-03 09:49:46 2.042121172 secs 1.110816e-02 secs
#> 41        resolved 2022-06-03 09:49:46 2.064169407 secs 1.099920e-02 secs
#> 42        resolved 2022-06-03 09:49:46 2.085657358 secs 1.100087e-02 secs
#> 43        resolved 2022-06-03 09:49:46 2.107613564 secs 1.089025e-02 secs
#> 44        resolved 2022-06-03 09:49:46 2.129410267 secs 1.098800e-02 secs
#> 45        resolved 2022-06-03 09:49:46 2.151370049 secs 1.101446e-02 secs
#> 46        resolved 2022-06-03 09:49:46 2.173346996 secs 1.093626e-02 secs
#> 47        resolved 2022-06-03 09:49:46 2.194980383 secs 1.096392e-02 secs
#> 48        resolved 2022-06-03 09:49:46 2.216877460 secs 1.094890e-02 secs
#> 49        resolved 2022-06-03 09:49:46 2.238571644 secs 1.118207e-02 secs
#> 50        resolved 2022-06-03 09:49:46 2.260672569 secs 1.099896e-02 secs
#> 51        resolved 2022-06-03 09:49:46 2.282336712 secs 1.093745e-02 secs
#> 52        resolved 2022-06-03 09:49:46 2.304240227 secs 1.094508e-02 secs
#> 53        resolved 2022-06-03 09:49:46 2.325916290 secs 1.095033e-02 secs
#> 54        resolved 2022-06-03 09:49:46 2.347764492 secs 1.055312e-02 secs
#> 55        resolved 2022-06-03 09:49:46 2.369407177 secs 1.123476e-02 secs
#> 56        resolved 2022-06-03 09:49:46 2.391613007 secs 1.075745e-02 secs
#> 57        resolved 2022-06-03 09:49:46 2.413007021 secs 1.060367e-02 secs
#> 58        resolved 2022-06-03 09:49:46 2.434431553 secs 1.093197e-02 secs
#> 59        resolved 2022-06-03 09:49:46 2.456258297 secs 1.092458e-02 secs
#> 60        resolved 2022-06-03 09:49:46 2.477954626 secs 1.091528e-02 secs
#> 61        resolved 2022-06-03 09:49:46 2.499757051 secs 1.091719e-02 secs
#> 62        resolved 2022-06-03 09:49:46 2.521594763 secs 1.058006e-02 secs
#> 63        resolved 2022-06-03 09:49:46 2.543099642 secs 1.084280e-02 secs
#> 64        resolved 2022-06-03 09:49:46 2.564869165 secs 1.090908e-02 secs
#> 65        resolved 2022-06-03 09:49:46 2.586640358 secs 1.095390e-02 secs
#> 66        resolved 2022-06-03 09:49:46 2.608518124 secs 1.095843e-02 secs
#> 67        resolved 2022-06-03 09:49:46 2.629990816 secs 1.099420e-02 secs
#> 68        resolved 2022-06-03 09:49:46 2.651947498 secs 1.074219e-02 secs
#> 69        resolved 2022-06-03 09:49:46 2.673363924 secs 1.128316e-02 secs
#> 70        resolved 2022-06-03 09:49:46 2.695616722 secs 1.098347e-02 secs
#> 71        resolved 2022-06-03 09:49:46 2.717518330 secs 1.094270e-02 secs
#> 72        resolved 2022-06-03 09:49:46 2.739350319 secs 1.064563e-02 secs
#> 73        resolved 2022-06-03 09:49:46 2.760881662 secs 1.078486e-02 secs
#> 74        resolved 2022-06-03 09:49:46 2.782604218 secs 1.100373e-02 secs
#> 75        resolved 2022-06-03 09:49:46 2.804512262 secs 1.096535e-02 secs
#> 76        resolved 2022-06-03 09:49:46 2.826364756 secs 1.071668e-02 secs
#> 77        resolved 2022-06-03 09:49:46 2.848271370 secs 1.113915e-02 secs
#> 78        resolved 2022-06-03 09:49:46 2.869965553 secs 1.100802e-02 secs
#> 79        resolved 2022-06-03 09:49:46 2.891873121 secs 1.085830e-02 secs
#> 80        resolved 2022-06-03 09:49:46 2.913615465 secs 1.098490e-02 secs
#> 81        resolved 2022-06-03 09:49:46 2.935499430 secs 1.059747e-02 secs
#> 82        resolved 2022-06-03 09:49:46 2.956994534 secs 1.093483e-02 secs
#> 83        resolved 2022-06-03 09:49:47 2.978848934 secs 1.099110e-02 secs
#> 84        resolved 2022-06-03 09:49:47 3.000394106 secs 1.074028e-02 secs
#> 85        resolved 2022-06-03 09:49:47 3.022026539 secs 1.092505e-02 secs
#> 86        resolved 2022-06-03 09:49:47 3.043546915 secs 1.143146e-02 secs
#> 87        resolved 2022-06-03 09:49:47 3.065957308 secs 1.412272e-02 secs
#> 88        resolved 2022-06-03 09:49:47 3.090970516 secs 1.097155e-02 secs
#> 89        resolved 2022-06-03 09:49:47 3.112832785 secs 1.091838e-02 secs
#> 90        resolved 2022-06-03 09:49:47 3.134625435 secs 1.094747e-02 secs
#> 91        resolved 2022-06-03 09:49:47 3.156525612 secs 1.111603e-02 secs
#> 92        resolved 2022-06-03 09:49:47 3.178275824 secs 1.136422e-02 secs
#> 93        resolved 2022-06-03 09:49:47 3.200514317 secs 1.099443e-02 secs
#> 94        resolved 2022-06-03 09:49:47 3.222002506 secs 1.092482e-02 secs
#> 95        resolved 2022-06-03 09:49:47 3.243809223 secs 1.087761e-02 secs
#> 96        resolved 2022-06-03 09:49:47 3.265216351 secs 1.093912e-02 secs
#> 97        resolved 2022-06-03 09:49:47 3.287197113 secs 1.094174e-02 secs
#> 98        resolved 2022-06-03 09:49:47 3.309022903 secs 1.059628e-02 secs
#> 99        resolved 2022-06-03 09:49:47 3.330459118 secs 1.074195e-02 secs
#> 100       resolved 2022-06-03 09:49:47 3.351737499 secs 1.056314e-02 secs
#> 101       resolved 2022-06-03 09:49:47 3.372806549 secs 1.055980e-02 secs
#> 102       resolved 2022-06-03 09:49:47 3.393879175 secs 1.056123e-02 secs
#> 103       resolved 2022-06-03 09:49:47 3.414962292 secs 1.056242e-02 secs
#> 104       resolved 2022-06-03 09:49:47 3.436100960 secs 1.053405e-02 secs
#> 105       resolved 2022-06-03 09:49:47 3.457145691 secs 1.056743e-02 secs
#> 106       resolved 2022-06-03 09:49:47 3.478222847 secs 1.050377e-02 secs
#> 107       resolved 2022-06-03 09:49:47 3.499255180 secs 1.057506e-02 secs
#> 108       resolved 2022-06-03 09:49:47 3.520645618 secs 1.095223e-02 secs
#> 109       resolved 2022-06-03 09:49:47 3.542490721 secs 1.093316e-02 secs
#> 110       resolved 2022-06-03 09:49:47 3.564317226 secs 1.092815e-02 secs
#> 111       resolved 2022-06-03 09:49:47 3.585954189 secs 1.103258e-02 secs
#> 112       resolved 2022-06-03 09:49:47 3.607887268 secs 1.106715e-02 secs
#> 113       resolved 2022-06-03 09:49:47 3.629638910 secs 1.124382e-02 secs
#> 114       resolved 2022-06-03 09:49:47 3.651984930 secs 1.110458e-02 secs
#> 115       resolved 2022-06-03 09:49:47 3.673922539 secs 1.098585e-02 secs
#> 116       resolved 2022-06-03 09:49:47 3.695806265 secs 1.105452e-02 secs
#> 117       resolved 2022-06-03 09:49:47 3.717675209 secs 1.087260e-02 secs
#> 118       resolved 2022-06-03 09:49:47 3.739076138 secs 1.050186e-02 secs
#> 119       resolved 2022-06-03 09:49:47 3.760104418 secs 1.096582e-02 secs
#> 120       resolved 2022-06-03 09:49:47 3.781931877 secs 1.057911e-02 secs
#> 121       resolved 2022-06-03 09:49:47 3.803420067 secs 1.063800e-02 secs
#> 122       resolved 2022-06-03 09:49:47 3.824797153 secs 1.073909e-02 secs
#> 123       resolved 2022-06-03 09:49:47 3.846576452 secs 1.104259e-02 secs
#> 124       resolved 2022-06-03 09:49:47 3.868307352 secs 1.097345e-02 secs
#> 125       resolved 2022-06-03 09:49:47 3.889955521 secs 1.092482e-02 secs
#> 126       resolved 2022-06-03 09:49:47 3.911768436 secs 1.095462e-02 secs
#> 127       resolved 2022-06-03 09:49:47 3.933639526 secs 1.102591e-02 secs
#> 128       resolved 2022-06-03 09:49:47 3.955266714 secs 1.071382e-02 secs
#> 131       resolved 2022-06-03 09:49:48 3.976599693 secs 8.493185e-03 secs
#> 129         gather 2022-06-03 09:49:48 3.983328342 secs 3.767014e-05 secs
#> 132        resolve 2022-06-03 09:49:48 3.985828400 secs 1.072884e-05 secs
#>     future_label                          future_uuid
#> 1           <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 6           <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 2           <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 3           <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 4           <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 5           <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 7           <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 130         <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 8           <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 9           <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 10          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 11          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 12          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 13          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 14          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 15          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 16          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 17          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 18          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 19          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 20          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 21          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 22          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 23          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 24          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 25          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 26          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 27          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 28          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 29          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 30          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 31          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 32          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 33          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 34          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 35          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 36          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 37          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 38          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 39          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 40          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 41          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 42          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 43          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 44          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 45          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 46          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 47          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 48          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 49          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 50          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 51          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 52          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 53          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 54          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 55          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 56          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 57          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 58          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 59          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 60          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 61          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 62          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 63          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 64          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 65          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 66          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 67          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 68          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 69          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 70          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 71          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 72          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 73          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 74          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 75          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 76          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 77          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 78          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 79          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 80          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 81          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 82          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 83          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 84          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 85          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 86          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 87          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 88          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 89          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 90          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 91          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 92          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 93          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 94          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 95          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 96          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 97          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 98          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 99          <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 100         <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 101         <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 102         <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 103         <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 104         <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 105         <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 106         <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 107         <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 108         <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 109         <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 110         <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 111         <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 112         <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 113         <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 114         <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 115         <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 116         <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 117         <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 118         <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 119         <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 120         <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 121         <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 122         <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 123         <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 124         <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 125         <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 126         <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 127         <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 128         <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 131         <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 129         <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#> 132         <NA> 55bfab9e-0f19-f06a-cf99-dcce3c22a6b4
#>                             session_uuid
#> 1   4575a0a0-1e24-f629-72bb-62e514a092c5
#> 6   4575a0a0-1e24-f629-72bb-62e514a092c5
#> 2   4575a0a0-1e24-f629-72bb-62e514a092c5
#> 3   4575a0a0-1e24-f629-72bb-62e514a092c5
#> 4   4575a0a0-1e24-f629-72bb-62e514a092c5
#> 5   4575a0a0-1e24-f629-72bb-62e514a092c5
#> 7   4575a0a0-1e24-f629-72bb-62e514a092c5
#> 130 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 8   4575a0a0-1e24-f629-72bb-62e514a092c5
#> 9   4575a0a0-1e24-f629-72bb-62e514a092c5
#> 10  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 11  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 12  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 13  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 14  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 15  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 16  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 17  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 18  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 19  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 20  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 21  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 22  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 23  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 24  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 25  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 26  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 27  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 28  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 29  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 30  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 31  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 32  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 33  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 34  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 35  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 36  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 37  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 38  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 39  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 40  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 41  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 42  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 43  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 44  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 45  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 46  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 47  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 48  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 49  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 50  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 51  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 52  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 53  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 54  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 55  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 56  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 57  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 58  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 59  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 60  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 61  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 62  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 63  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 64  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 65  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 66  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 67  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 68  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 69  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 70  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 71  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 72  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 73  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 74  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 75  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 76  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 77  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 78  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 79  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 80  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 81  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 82  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 83  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 84  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 85  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 86  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 87  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 88  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 89  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 90  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 91  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 92  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 93  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 94  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 95  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 96  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 97  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 98  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 99  4575a0a0-1e24-f629-72bb-62e514a092c5
#> 100 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 101 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 102 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 103 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 104 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 105 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 106 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 107 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 108 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 109 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 110 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 111 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 112 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 113 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 114 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 115 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 116 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 117 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 118 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 119 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 120 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 121 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 122 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 123 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 124 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 125 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 126 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 127 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 128 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 131 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 129 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 132 4575a0a0-1e24-f629-72bb-62e514a092c5
#> 

## Stop parallel workers and undo journaling settings
plan(sequential)
options(oopts)