make_rng_seeds {future}	R Documentation

Produce Reproducible Seeds for Parallel Random Number Generation

Description

Produce Reproducible Seeds for Parallel Random Number Generation

Usage

make_rng_seeds(count, seed = FALSE, debug = getOption("future.debug", FALSE))

Arguments

count
The number of RNG seeds to produce.

seed
A logical specifying whether RNG seeds should be generated or not. (seed = NULL corresponds to seed = FALSE). If a list, then it should be of length count and each element should consist of a valid RNG seed.

debug
If TRUE, debug output is produced, otherwise not.

Value

Returns a non-named list of length count, or NULL. Any seed returned is a valid RNG seed.

[Package future version 1.21.0 Index]