re-exports {future}	R Documentation

Functions Moved to 'parallelly'

Description

The following function used to be part of future, but has since been migrated to parallelly. The migration started with future 1.20.0 (November 2020). They were moved because they are also useful outside of the future framework.

Details

If you are using any of these from the future package, please switch to use the ones from the parallelly package. Thank you!

parallelly::as.cluster() (no longer re-exported)

parallelly::autoStopCluster() (no longer re-exported)

parallelly::availableCores()

parallelly::availableWorkers()

parallelly::makeClusterMPI() (no longer re-exported)

parallelly::makeClusterPSOCK()

parallelly::makeNodePSOCK() (no longer re-exported)

parallelly::supportsMulticore()

For backward-compatible reasons, some of these functions remain available as exact copies also from this package (as re-exports), e.g.

cl <- parallelly::makeClusterPSOCK(2)

can still be accessed as:

cl <- future::makeClusterPSOCK(2)

Note that it is the goal to remove all of the above from this package.

[Package future version 1.70.0 Index]