result.Future {future}	R Documentation

Get the results of a resolved future

Description

Get the results of a resolved future

Usage

## S3 method for class 'Future'
result(future, ...)

Arguments

future
A Future.

...
Not used.

Details

This function is only part of the backend Future API. This function is not part of the frontend Future API.

Value

The FutureResult object. It may signal a FutureError, if there is a significant orchestration error. For example, if the parallel worker process terminated abruptly ("crashed"), then a FutureInterruptError is signaled.

[Package future version 1.70.0 Index]