run.Future {future}	R Documentation

Run a future

Description

Run a future

Usage

## S3 method for class 'Future'
run(future, ...)

Arguments

future
A Future.

...
Not used.

Details

This function can only be called once per future. Further calls will result in an informative error. If a future is not run when its value is queried, then it is run at that point.

Value

The Future object.

[Package future version 1.21.0 Index]