save_rds {future}	R Documentation

Robustly Saves an Object to RDS File Atomically

Description

Robustly Saves an Object to RDS File Atomically

Usage

save_rds(object, pathname, ...)

Arguments

object
The R object to be save.

pathname
RDS file to written.

...
(optional) Additional arguments passed to base::saveRDS().

Details

Uses base::saveRDS internally but writes the object atomically by first writing to a temporary file which is then renamed.

Value

(invisible) The pathname of the RDS written.

[Package future version 1.21.0 Index]