signalConditions {future}	R Documentation

Signals Captured Conditions

Description

Captured conditions that meet the include and exclude requirements are signaled in the order as they were captured.

Usage

signalConditions(
  future,
  include = "condition",
  exclude = NULL,
  resignal = TRUE,
  ...
)

Arguments

future
A resolved Future.

include
A character string of condition classes to signal.

exclude
A character string of condition classes not to signal.

resignal
If TRUE, then already signaled conditions are signaled again, otherwise not.

...
Not used.

Value

Returns the Future where conditioned that were signaled have been flagged to have been signaled.

See Also

Conditions are signaled by signalCondition().

[Package future version 1.70.0 Index]