sticky_globals {future}	R Documentation

Place a sticky-globals environment immediately after the global environment

Description

Place a sticky-globals environment immediately after the global environment

Usage

sticky_globals(erase = FALSE, name = "future:sticky_globals", pos = 2L)

Arguments

erase
(logical) If TRUE, the environment is erased, otherwise not.

name
(character) The name of the environment on the base::search path.

pos
(integer) The position on the search path where the environment should be positioned. If pos == 0L, then the environment is detached, if it exists.

Value

(invisible; environment) The environment.

[Package future version 1.21.0 Index]