tweak {future}	R Documentation

Tweak a future function by adjusting its default arguments

Description

Tweak a future function by adjusting its default arguments

Usage

tweak(strategy, ..., penvir = parent.frame())

Arguments

strategy
An existing future function or the name of one.

...
Named arguments to replace the defaults of existing arguments.

penvir
The environment used when searching for a future function by its name.

Value

a future function.

See Also

Use plan() to set a future to become the new default strategy.

[Package future version 1.21.0 Index]