usedCores {future}	R Documentation

Get number of cores currently used

Description

Get number of children (and don't count the current process) used by the current R session. The number of children is the total number of subprocesses launched by this process that are still running and whose values have yet not been collected.

Usage

usedCores()

Value

A non-negative integer.

[Package future version 1.21.0 Index]