ClusterFuture-class {future}	R Documentation

A cluster future is a future whose value will be resolved asynchronously in a parallel process

Description

A cluster future is a future whose value will be resolved asynchronously in a parallel process

Usage

ClusterFuture(
  expr = NULL,
  substitute = TRUE,
  envir = parent.frame(),
  globals = TRUE,
  packages = NULL,
  local = !persistent,
  persistent = FALSE,
  workers = NULL,
  user = NULL,
  master = NULL,
  revtunnel = TRUE,
  homogeneous = TRUE,
  ...
)

MultisessionFuture(
  expr = NULL,
  substitute = TRUE,
  envir = parent.frame(),
  persistent = FALSE,
  workers = NULL,
  ...
)

Arguments

expr
An R expression.

substitute
If TRUE, argument expr is substitute():ed, otherwise not.

envir
The environment from where global objects should be identified.

globals
(optional) a logical, a character vector, or a named list to control how globals are handled. For details, see section 'Globals used by future expressions' in the help for future().

packages
(optional) a character vector specifying packages to be attached in the R environment evaluating the future.

local
If TRUE, the expression is evaluated such that all assignments are done to local temporary environment, otherwise the assignments are done to the global environment of the R process evaluating the future.

persistent
If FALSE, the evaluation environment is cleared from objects prior to the evaluation of the future.

workers
A cluster object, a character vector of host names, a positive numeric scalar, or a function. If a character vector or a numeric scalar, a cluster object is created using makeClusterPSOCK(workers). If a function, it is called without arguments when the future is created and its value is used to configure the workers. The function should return any of the above types.

user
(optional) The user name to be used when communicating with another host.

master
(optional) The hostname or IP address of the master machine running this node.

revtunnel
If TRUE, reverse SSH tunneling is used for the PSOCK cluster nodes to connect back to the master R process. This avoids the hassle of firewalls, port forwarding and having to know the internal / public IP address of the master R session.

homogeneous
If TRUE, all cluster nodes is assumed to use the same path to ‘Rscript’ as the main R session. If FALSE, the it is assumed to be on the PATH for each node.

...
Additional named elements passed to Future().

Value

ClusterFuture() returns an object of class ClusterFuture.

MultisessionFuture() returns an object of class MultisessionFuture, which inherits from ClusterFuture.

See Also

To evaluate an expression using "cluster future", see function cluster().

[Package future version 1.21.0 Index]