ConstantFuture-class {future}	R Documentation

A future with a constant value

Description

A constant future is a future whose expression is a constant and therefore by definition is already resolved upon creation.

Usage

ConstantFuture(..., globals = TRUE, envir = emptyenv())

Arguments

...
Additional named elements of the future.

globals
(optional) a logical, a character vector, or a named list to control how globals are handled. For details, see section 'Globals used by future expressions' in the help for future().

envir
The environment from where global objects should be identified.

Value

An object of class ConstantFuture.

[Package future version 1.21.0 Index]