Future-class {future}	R Documentation

A future represents a value that will be available at some point in the future

Description

A future is an abstraction for a value that may available at some point in the future. A future can either be unresolved or resolved, a state which can be checked with resolved(). As long as it is unresolved, the value is not available. As soon as it is resolved, the value is available via value().

Usage

Future(
  expr = NULL,
  envir = parent.frame(),
  substitute = TRUE,
  stdout = TRUE,
  conditions = "condition",
  globals = NULL,
  packages = NULL,
  seed = FALSE,
  lazy = FALSE,
  local = TRUE,
  gc = FALSE,
  earlySignal = FALSE,
  label = NULL,
  ...
)

Arguments

expr
An R expression.

envir
The environment from where global objects should be identified.

substitute
If TRUE, argument expr is substitute():ed, otherwise not.

stdout
If TRUE (default), then the standard output is captured, and re-outputted when value() is called. If FALSE, any output is silenced (by sinking it to the null device as it is outputted). If NA (not recommended), output is not intercepted.

conditions
A character string of conditions classes to be captured and relayed. The default is to relay messages and warnings. To not intercept any types of conditions, use conditions = NULL. Errors are always relayed.

globals
(optional) a logical, a character vector, or a named list to control how globals are handled. For details, see section 'Globals used by future expressions' in the help for future().

packages
(optional) a character vector specifying packages to be attached in the R environment evaluating the future.

seed
(optional) If TRUE, the random seed, that is, the state of the random number generator (RNG) will be set such that statistically sound random numbers are produced (also during parallelization). If FALSE (default), it is assumed that the future expression does neither need nor use random numbers generation. To use a fixed random seed, specify a L'Ecuyer-CMRG seed (seven integer) or a regular RNG seed (a single integer). If the latter, then a L'Ecuyer-CMRG seed will be automatically created based on the given seed. Furthermore, if FALSE, then the future will be monitored to make sure it does not use random numbers. If it does and depending on the value of option future.rng.onMisuse, the check is ignored, an informative warning, or error will be produced. If seed is NULL, then the effect is as with seed = FALSE but without the RNG check being performed.

lazy
If FALSE (default), the future is resolved eagerly (starting immediately), otherwise not.

local
If TRUE, the expression is evaluated such that all assignments are done to local temporary environment, otherwise the assignments are done to the global environment of the R process evaluating the future.

gc
If TRUE, the garbage collector run (in the process that evaluated the future) only after the value of the future is collected. Exactly when the values are collected may depend on various factors such as number of free workers and whether earlySignal is TRUE (more frequently) or FALSE (less frequently). Some types of futures ignore this argument.

earlySignal
Specified whether conditions should be signaled as soon as possible or not.

label
An optional character string label attached to the future.

...
Additional named elements of the future.

Details

A Future object is itself an environment.

Value

An object of class Future.

See Also

One function that creates a Future is future(). It returns a Future that evaluates an R expression in the future. An alternative approach is to use the %<-% infix assignment operator, which creates a future from the right-hand-side (RHS) R expression and assigns its future value to a variable as a promise.

[Package future version 1.21.0 Index]