FutureCondition {future}	R Documentation

A condition (message, warning, or error) that occurred while orchestrating a future

Description

While orchestrating (creating, launching, querying, collection) futures, unexpected run-time errors (and other types of conditions) may occur. Such conditions are coerced to a corresponding FutureCondition class to help distinguish them from conditions that occur due to the evaluation of the future.

Usage

FutureCondition(message, call = NULL, future = NULL)

FutureMessage(message, call = NULL, future = NULL)

FutureWarning(message, call = NULL, future = NULL)

FutureError(message, call = NULL, future = NULL)

RngFutureCondition(message = NULL, call = NULL, future)

RngFutureWarning(...)

RngFutureError(...)

UnexpectedFutureResultError(future, hint = NULL)

Arguments

message
A message condition.

call
The call stack that led up to the condition.

future
The Future involved.

hint
(optional) A string with a suggestion on what might be wrong.

Value

An object of class FutureCondition which inherits from class condition and FutureMessage, FutureWarning, and FutureError all inherits from FutureCondition. Moreover, a FutureError inherits from error, a FutureWarning from warning, and a FutureMessage from message.

[Package future version 1.21.0 Index]