MultiprocessFuture-class {future}	R Documentation

An multiprocess future is a future whose value will be resolved asynchronously in a parallel process

Description

An multiprocess future is a future whose value will be resolved asynchronously in a parallel process

Usage

MultiprocessFuture(expr = NULL, substitute = TRUE, envir = parent.frame(), ...)

Arguments

expr
An R expression.

substitute
If TRUE, argument expr is substitute():ed, otherwise not.

envir
The environment from where global objects should be identified.

...
Additional named elements passed to Future().

Value

An object of class MultiprocessFuture.

See Also

To evaluate an expression using "multiprocess future", see function multiprocess().

[Package future version 1.21.0 Index]