UniprocessFuture-class {future}	R Documentation

An uniprocess future is a future whose value will be resolved synchronously in the current process

Description

An uniprocess future is a future whose value will be resolved synchronously in the current process

Usage

UniprocessFuture(
  expr = NULL,
  substitute = TRUE,
  envir = parent.frame(),
  globals = TRUE,
  packages = NULL,
  lazy = FALSE,
  local = TRUE,
  ...
)

SequentialFuture(
  expr = NULL,
  envir = parent.frame(),
  substitute = TRUE,
  lazy = FALSE,
  globals = TRUE,
  local = TRUE,
  ...
)

Arguments

expr
An R expression.

substitute
If TRUE, argument expr is substitute():ed, otherwise not.

envir
The environment from where global objects should be identified.

globals
(optional) a logical, a character vector, or a named list to control how globals are handled. For details, see section 'Globals used by future expressions' in the help for future().

packages
(optional) a character vector specifying packages to be attached in the R environment evaluating the future.

lazy
If FALSE (default), the future is resolved eagerly (starting immediately), otherwise not.

local
If TRUE, the expression is evaluated such that all assignments are done to local temporary environment, otherwise the assignments are done to the global environment of the R process evaluating the future.

...
Additional named elements passed to Future().

Value

An object of class UniprocessFuture.

See Also

To evaluate an expression using "uniprocess future", see functions uniprocess().

[Package future version 1.21.0 Index]