cluster {future}	R Documentation

Create a cluster future whose value will be resolved asynchronously in a parallel process

Description

A cluster future is a future that uses cluster evaluation, which means that its value is computed and resolved in parallel in another process.

Usage

cluster(..., workers = availableWorkers(), envir = parent.frame())

Arguments

...
Additional named elements passed to ClusterFuture().

workers
A cluster object, a character vector of host names, a positive numeric scalar, or a function. If a character vector or a numeric scalar, a cluster object is created using makeClusterPSOCK(workers). If a function, it is called without arguments when the future is created and its value is used to configure the workers. The function should return any of the above types.

envir
The environment from where global objects should be identified.

Details

This function will block if all available R cluster nodes are occupied and will be unblocked as soon as one of the already running cluster futures is resolved.

The preferred way to create an cluster future is not to call this function directly, but to register it via plan(cluster) such that it becomes the default mechanism for all futures. After this future() and %<-% will create cluster futures.

Value

A ClusterFuture.

Examples

## Use cluster futures
cl <- parallel::makeCluster(2, timeout = 60)
plan(cluster, workers = cl)

## A global variable
a <- 0

## Create future (explicitly)
f <- future({
  b <- 3
  c <- 2
  a * b * c
})

## A cluster future is evaluated in a separate process.
## Regardless, changing the value of a global variable will
## not affect the result of the future.
a <- 7
print(a)

v <- value(f)
print(v)
stopifnot(v == 0)

## CLEANUP
parallel::stopCluster(cl)

[Package future version 1.21.0 Index]