find_references {future}	R Documentation

Get first or all references of an R object

Description

Get first or all references of an R object

Assert that there are no references among the identified globals

Usage

find_references(x, first_only = FALSE)

assert_no_references(x, action = c("error", "warning", "message", "string"))

Arguments

x
The R object to be checked.

first_only
If TRUE, only the first reference is returned, otherwise all references.

action
Type of action to take if a reference is found.

Value

find_references() returns a list of one or more references identified.

If a reference is detected, an informative error, warning, message, or a character string is produced, otherwise NULL is returned invisibly.

[Package future version 1.21.0 Index]