Create a future

Creates a future that evaluates an R expression or a future that calls an R function with a set of arguments. How, when, and where these futures are evaluated can be configured using plan() such that it is evaluated in parallel on, for instance, the current machine, on a remote machine, or via a job queue on a compute cluster. Importantly, any R code using futures remains the same regardless on these settings and there is no need to modify the code when switching from, say, sequential to parallel processing.

future(
  expr,
  envir = parent.frame(),
  substitute = TRUE,
  lazy = FALSE,
  seed = FALSE,
  globals = TRUE,
  packages = NULL,
  stdout = TRUE,
  conditions = "condition",
  earlySignal = FALSE,
  label = NULL,
  gc = FALSE,
  ...
)

futureAssign(
  x,
  value,
  envir = parent.frame(),
  substitute = TRUE,
  lazy = FALSE,
  seed = FALSE,
  globals = TRUE,
  packages = NULL,
  stdout = TRUE,
  conditions = "condition",
  earlySignal = FALSE,
  label = NULL,
  gc = FALSE,
  ...,
  assign.env = envir
)

x %<-% value

futureCall(
  FUN,
  args = list(),
  envir = parent.frame(),
  lazy = FALSE,
  seed = FALSE,
  globals = TRUE,
  packages = NULL,
  stdout = TRUE,
  conditions = "condition",
  earlySignal = FALSE,
  label = NULL,
  gc = FALSE,
  ...
)

Arguments

expr, value
An R expression.

envir
The environment from where global objects should be identified.

substitute
If TRUE, argument expr is substitute():ed, otherwise not.

lazy
If FALSE (default), the future is resolved eagerly (starting immediately), otherwise not.

seed
(optional) If TRUE, the random seed, that is, the state of the random number generator (RNG) will be set such that statistically sound random numbers are produced (also during parallelization). If FALSE (default), it is assumed that the future expression does neither need nor use random numbers generation. To use a fixed random seed, specify a L'Ecuyer-CMRG seed (seven integer) or a regular RNG seed (a single integer). If the latter, then a L'Ecuyer-CMRG seed will be automatically created based on the given seed. Furthermore, if FALSE, then the future will be monitored to make sure it does not use random numbers. If it does and depending on the value of option future.rng.onMisuse, the check is ignored, an informative warning, or error will be produced. If seed is NULL, then the effect is as with seed = FALSE but without the RNG check being performed.

globals
(optional) a logical, a character vector, or a named list to control how globals are handled. For details, see section 'Globals used by future expressions' in the help for future().

packages
(optional) a character vector specifying packages to be attached in the R environment evaluating the future.

stdout
If TRUE (default), then the standard output is captured, and re-outputted when value() is called. If FALSE, any output is silenced (by sinking it to the null device as it is outputted). If NA (not recommended), output is not intercepted.

conditions
A character string of conditions classes to be captured and relayed. The default is to relay messages and warnings. To not intercept any types of conditions, use conditions = NULL. Attribute exclude can be used to ignore specific classes, e.g. conditions = structure("condition", exclude = "message") will capture all condition classes except those that inherits from the message class. Errors are always relayed.

earlySignal
Specified whether conditions should be signaled as soon as possible or not.

label
An optional character string label attached to the future.

gc
If TRUE, the garbage collector run (in the process that evaluated the future) only after the value of the future is collected. Exactly when the values are collected may depend on various factors such as number of free workers and whether earlySignal is TRUE (more frequently) or FALSE (less frequently). Some types of futures ignore this argument.

...
Additional arguments passed to Future().

x
the name of a future variable, which will hold the value of the future expression (as a promise).

assign.env
The environment to which the variable should be assigned.

FUN
A function to be evaluated.

args
A list of arguments passed to function FUN.

Value

f <- future(expr) creates a Future f that evaluates expression expr, the value of the future is retrieved using v <- value(f).

x %<-% value (a future assignment) and futureAssign("x", value) create a Future that evaluates expression expr and binds its value (as a promise) to a variable x. The value of the future is automatically retrieved when the assigned variable (promise) is queried. The future itself is returned invisibly, e.g. f <- futureAssign("x", expr) and f <- (x %<-% expr). Alternatively, the future of a future variable x can be retrieved without blocking using f <- futureOf(x). Both the future and the variable (promise) are assigned to environment assign.env where the name of the future is .future_<name>.

f <- futureCall(FUN, args) creates a Future f that calls function FUN with arguments args, where the value of the future is retrieved using x <- value(f).

Details

The state of a future is either unresolved or resolved. The value of a future can be retrieved using v <- value(f). Querying the value of a non-resolved future will block the call until the future is resolved. It is possible to check whether a future is resolved or not without blocking by using resolved(f).

For a future created via a future assignment (x %<-% value or futureAssign("x", value)), the value is bound to a promise, which when queried will internally call value() on the future and which will then be resolved into a regular variable bound to that value. For example, with future assignment x %<-% value, the first time variable x is queried the call blocks if (and only if) the future is not yet resolved. As soon as it is resolved, and any succeeding queries, querying x will immediately give the value.

The future assignment construct x %<-% value is not a formal assignment per se, but a binary infix operator on objects x and expression value. However, by using non-standard evaluation, this constructs can emulate an assignment operator similar to x <- value. Due to R's precedence rules of operators, future expressions often need to be explicitly bracketed, e.g. x %<-% { a + b }.

The futureCall() function works analogously to do.call(), which calls a function with a set of arguments. The difference is that do.call() returns the value of the call whereas futureCall() returns a future.

Eager or lazy evaluation

By default, a future is resolved using eager evaluation (lazy = FALSE). This means that the expression starts to be evaluated as soon as the future is created.

As an alternative, the future can be resolved using lazy evaluation (lazy = TRUE). This means that the expression will only be evaluated when the value of the future is requested. Note that this means that the expression may not be evaluated at all - it is guaranteed to be evaluated if the value is requested.

For future assignments, lazy evaluation can be controlled via the %lazy% operator, e.g. x %<-% { expr } %lazy% TRUE.

Globals used by future expressions

Global objects (short globals) are objects (e.g. variables and functions) that are needed in order for the future expression to be evaluated while not being local objects that are defined by the future expression. For example, in

  a <- 42
  f <- future({ b <- 2; a * b })

variable a is a global of future assignment f whereas b is a local variable. In order for the future to be resolved successfully (and correctly), all globals need to be gathered when the future is created such that they are available whenever and wherever the future is resolved.

The default behavior (globals = TRUE), is that globals are automatically identified and gathered. More precisely, globals are identified via code inspection of the future expression expr and their values are retrieved with environment envir as the starting point (basically via get(global, envir = envir, inherits = TRUE)). In most cases, such automatic collection of globals is sufficient and less tedious and error prone than if they are manually specified.

However, for full control, it is also possible to explicitly specify exactly which the globals are by providing their names as a character vector. In the above example, we could use

  a <- 42
  f <- future({ b <- 2; a * b }, globals = "a")

Yet another alternative is to explicitly specify also their values using a named list as in

  a <- 42
  f <- future({ b <- 2; a * b }, globals = list(a = a))

or

  f <- future({ b <- 2; a * b }, globals = list(a = 42))

Specifying globals explicitly avoids the overhead added from automatically identifying the globals and gathering their values. Furthermore, if we know that the future expression does not make use of any global variables, we can disable the automatic search for globals by using

  f <- future({ a <- 42; b <- 2; a * b }, globals = FALSE)

Future expressions often make use of functions from one or more packages. As long as these functions are part of the set of globals, the future package will make sure that those packages are attached when the future is resolved. Because there is no need for such globals to be frozen or exported, the future package will not export them, which reduces the amount of transferred objects. For example, in

  x <- rnorm(1000)
  f <- future({ median(x) })

variable x and median() are globals, but only x is exported whereas median(), which is part of the stats package, is not exported. Instead it is made sure that the stats package is on the search path when the future expression is evaluated. Effectively, the above becomes

  x <- rnorm(1000)
  f <- future({
    library("stats")
    median(x)
  })

To manually specify this, one can either do

  x <- rnorm(1000)
  f <- future({
    median(x)
  }, globals = list(x = x, median = stats::median)

or

  x <- rnorm(1000)
  f <- future({
    library("stats")
    median(x)
  }, globals = list(x = x))

Both are effectively the same.

Although rarely needed, a combination of automatic identification and manual specification of globals is supported via attributes add (to add false negatives) and ignore (to ignore false positives) on value TRUE. For example, with globals = structure(TRUE, ignore = "b", add = "a") any globals automatically identified except b will be used in addition to global a.

When using future assignments, globals can be specified analogously using the %globals% operator, e.g.

  x <- rnorm(1000)
  y %<-% { median(x) } %globals% list(x = x, median = stats::median)

See also

Author

The future logo was designed by Dan LaBar and tweaked by Henrik Bengtsson.

Examples

## Evaluate futures in parallel
plan(multisession)

## Data
x <- rnorm(100)
y <- 2 * x + 0.2 + rnorm(100)
w <- 1 + x ^ 2

## EXAMPLE: Regular assignments (evaluated sequentially)
fitA <- lm(y ~ x, weights = w)      ## with offset
fitB <- lm(y ~ x - 1, weights = w)  ## without offset
fitC <- {
  w <- 1 + abs(x)  ## Different weights
  lm(y ~ x, weights = w)
}
print(fitA)

#> 
#> Call:
#> lm(formula = y ~ x, weights = w)
#> 
#> Coefficients:
#> (Intercept)            x  
#>      0.2778       2.0244  
#> 

print(fitB)

#> 
#> Call:
#> lm(formula = y ~ x - 1, weights = w)
#> 
#> Coefficients:
#>     x  
#> 2.028  
#> 

print(fitC)

#> 
#> Call:
#> lm(formula = y ~ x, weights = w)
#> 
#> Coefficients:
#> (Intercept)            x  
#>      0.2938       1.9952  
#> 

## EXAMPLE: Future assignments (evaluated in parallel)
fitA %<-% lm(y ~ x, weights = w)      ## with offset
fitB %<-% lm(y ~ x - 1, weights = w)  ## without offset
fitC %<-% {
  w <- 1 + abs(x)
  lm(y ~ x, weights = w)
}
print(fitA)

#> 
#> Call:
#> lm(formula = y ~ x, weights = w)
#> 
#> Coefficients:
#> (Intercept)            x  
#>      0.2938       1.9952  
#> 

print(fitB)

#> 
#> Call:
#> lm(formula = y ~ x - 1, weights = w)
#> 
#> Coefficients:
#>     x  
#> 2.005  
#> 

print(fitC)

#> 
#> Call:
#> lm(formula = y ~ x, weights = w)
#> 
#> Coefficients:
#> (Intercept)            x  
#>      0.2938       1.9952  
#> 

## EXAMPLE: Explicitly create futures (evaluated in parallel)
## and retrieve their values
fA <- future( lm(y ~ x, weights = w) )
fB <- future( lm(y ~ x - 1, weights = w) )
fC <- future({
  w <- 1 + abs(x)
  lm(y ~ x, weights = w)
})
fitA <- value(fA)
fitB <- value(fB)
fitC <- value(fC)
print(fitA)

#> 
#> Call:
#> lm(formula = y ~ x, weights = w)
#> 
#> Coefficients:
#> (Intercept)            x  
#>      0.2938       1.9952  
#> 

print(fitB)

#> 
#> Call:
#> lm(formula = y ~ x - 1, weights = w)
#> 
#> Coefficients:
#>     x  
#> 2.005  
#> 

print(fitC)

#> 
#> Call:
#> lm(formula = y ~ x, weights = w)
#> 
#> Coefficients:
#> (Intercept)            x  
#>      0.2938       1.9952  
#> 

# \dontshow{
## Make sure to "close" an multisession workers on Windows
plan(sequential)
# }
## EXAMPLE: futureCall() and do.call()
x <- 1:100
y0 <- do.call(sum, args = list(x))
print(y0)

#> [1] 5050

f1 <- futureCall(sum, args = list(x))
y1 <- value(f1)
print(y1)

#> [1] 5050