re-exports {future}	R Documentation

Functions Moved to 'parallelly'

Description

The following function used to be part of future but has since been migrated to parallelly. The migration started with future 1.20.0 (November 2020). They were moved because they are also useful outside of the future framework.

Details

parallelly::as.cluster()

parallelly::autoStopCluster()

parallelly::availableCores()

parallelly::availableWorkers()

parallelly::makeClusterMPI()

parallelly::makeClusterPSOCK()

parallelly::makeNodePSOCK()

parallelly::supportsMulticore()

For backward-compatible reasons, these functions remain available as exact copies also from this package (as re-exports). For example,

cl <- parallelly::makeClusterPSOCK(2)

can still be accessed as:

cl <- future::makeClusterPSOCK(2)

[Package future version 1.21.0 Index]