remote {future}	R Documentation

Create a remote future whose value will be resolved asynchronously in a remote process

Description

A remote future is a future that uses remote cluster evaluation, which means that its value is computed and resolved remotely in another process.

Usage

remote(
  ...,
  workers = NULL,
  revtunnel = TRUE,
  myip = NULL,
  persistent = TRUE,
  envir = parent.frame()
)

Arguments

...
Additional named elements passed to Future().

workers
A cluster object, a character vector of host names, a positive numeric scalar, or a function. If a character vector or a numeric scalar, a cluster object is created using makeClusterPSOCK(workers). If a function, it is called without arguments when the future is created and its value is used to configure the workers. The function should return any of the above types.

revtunnel
If TRUE, reverse SSH tunneling is used for the PSOCK cluster nodes to connect back to the master R process. This avoids the hassle of firewalls, port forwarding and having to know the internal / public IP address of the master R session.

myip
The external IP address of this machine. If NULL, then it is inferred using an online service (default).

persistent
If FALSE, the evaluation environment is cleared from objects prior to the evaluation of the future.

envir
The environment from where global objects should be identified.

Value

A ClusterFuture.

'remote' versus 'cluster'

The remote plan is a very similar to the cluster plan, but provides more convenient default argument values when connecting to remote machines. Specifically, remote uses persistent = TRUE by default, and it sets homogeneous, revtunnel, and myip "wisely" depending on the value of workers. See below for example on how remote and cluster are related.

Examples

## Not run: \donttest{

## Use a remote machine
plan(remote, workers = "remote.server.org")

## Evaluate expression remotely
host %<-% { Sys.info()[["nodename"]] }
host
[1] "remote.server.org"

## The following setups are equivalent:
plan(remote, workers = "localhost")
plan(cluster, workers = "localhost", persistent = TRUE)
plan(cluster, workers = 1, persistent = TRUE)
plan(multisession, workers = 1, persistent = TRUE)

## The following setups are equivalent:
plan(remote, workers = "remote.server.org")
plan(cluster, workers = "remote.server.org", persistent = TRUE, homogeneous = FALSE)

## The following setups are equivalent:
cl <- makeClusterPSOCK("remote.server.org")
plan(remote, workers = cl)
plan(cluster, workers = cl, persistent = TRUE)

}
## End(Not run)

[Package future version 1.21.0 Index]