resolved {future}	R Documentation

Check whether a future is resolved or not

Description

Check whether a future is resolved or not

Usage

resolved(x, ...)

Arguments

x
A Future, a list, or an environment (which also includes list environment.

...
Not used.

Details

This method needs to be implemented by the class that implement the Future API. The implementation should return either TRUE or FALSE and must never throw an error (except for FutureError:s which indicate significant, often unrecoverable infrastructure problems). It should also be possible to use the method for polling the future until it is resolved (without having to wait infinitely long), e.g. while (!resolved(future)) Sys.sleep(5).

Value

A logical of the same length and dimensions as x. Each element is TRUE unless the corresponding element is a non-resolved future in case it is FALSE.

[Package future version 1.21.0 Index]