result.Future {future}	R Documentation

Get the results of a resolved future

Description

Get the results of a resolved future

Usage

## S3 method for class 'Future'
result(future, ...)

Arguments

future
A Future.

...
Not used.

Details

This function is only part of the backend Future API. This function is not part of the frontend Future API.

Value

The FutureResult object.

[Package future version 1.21.0 Index]