sequential {future}	R Documentation

Create a sequential future whose value will be in the current R session

Description

A sequential future is a future that is evaluated sequentially in the current R session similarly to how R expressions are evaluated in R. The only difference to R itself is that globals are validated by default just as for all other types of futures in this package.

Usage

sequential(..., envir = parent.frame())

transparent(..., envir = parent.frame())

Arguments

...
Additional arguments passed to Future().

envir
The environment from where global objects should be identified.

Details

The preferred way to create a sequential future is not to call these functions directly, but to register them via plan(sequential) such that it becomes the default mechanism for all futures. After this future() and %<-% will create sequential futures.

Value

A SequentialFuture.

transparent futures

Transparent futures are sequential futures configured to emulate how R evaluates expressions as far as possible. For instance, errors and warnings are signaled immediately and assignments are done to the calling environment (without local() as default for all other types of futures). This makes transparent futures ideal for troubleshooting, especially when there are errors.

Examples

## Use sequential futures
plan(sequential)

## A global variable
a <- 0

## Create a sequential future
f <- future({
  b <- 3
  c <- 2
  a * b * c
})

## Since 'a' is a global variable in future 'f' which
## is eagerly resolved (default), this global has already
## been resolved / incorporated, and any changes to 'a'
## at this point will _not_ affect the value of 'f'.
a <- 7
print(a)

v <- value(f)
print(v)
stopifnot(v == 0)

[Package future version 1.21.0 Index]